[IMP]: routes and code refactoring
