+++ app/Http/Controllers/PatientController.php
@@ -4,23 +4,22 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\patient\PatientUpdateRequest;
 use App\Http\Requests\PatientStoreRequest;
-use App\Models\Appointment;
 use App\Models\Patient;
 use App\Services\PatientService;
-use DateTime;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Storage;
 
 class PatientController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $patients = Patient::paginate(10);
-        return view('patient.index', ['patients' => $patients]);
+        $patientService = new PatientService();
+        $data =  $patientService->show($request->input('search'));
+        $data['currentPage'] = $request->input('page', 1);
+        return view('patient.index', $data);
     }
 
     /**
@@ -72,41 +71,6 @@ class PatientController extends Controller
         return view('patient.edit', compact('patient'));
     }
 
-    public function adminPatientUpdate(Patient $patient, Request $request)
-    {
-        // Validate the incoming request data
-        $validatedData = $request->validate([
-            'name' => 'required|string|max:255', // Added max length for name
-            'dob' => 'required|date|before:today', // Ensure DOB is before today
-            'gender' => 'required|in:male,female,other',
-            'image' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
-        ]);
-
-        $patient->user->name = $validatedData['name'];
-
-        $patient->dob = $validatedData['dob'];
-        $patient->gender = $validatedData['gender'];
-
-        // Handle file upload
-        if ($request->hasFile('image')) {
-            if ($patient->image && Storage::exists('public/uploads_patient/' . $patient->image)) {
-                Storage::delete('public/uploads_patient/' . $patient->image);
-            }
-
-            $imagePath = $request->file('image')->store('uploads_patient', 'public');
-            $patient->image = basename($imagePath);
-        }
-
-        $patient->user->save();
-
-        $patient->save();
-
-        // Redirect back with success message
-        return redirect()->back()->with('status', [
-            'message' => 'Patient updated successfully',
-            'type' => 'success'
-        ]);
-    }
 
     /**
      * Update the specified resource in storage.
@@ -132,21 +96,5 @@ class PatientController extends Controller
             'message' =>  'Patient ' . $patient->user->name .  ' has been deleted successfully.',
             'type' => 'failure'
         ]);
-    }
-
-    public function search(Request $request)
-    {
-        $searchTerm = $request->input('search');
-
-        $patients = Patient::with('user')
-            ->where('gender', 'like', "%{$searchTerm}%")
-            ->orWhereHas('user', function ($query) use ($searchTerm) {
-                $query->where('name', 'like', "%{$searchTerm}%")
-                    ->orWhere('email', 'like', "%{$searchTerm}%");
-            })
-            ->get();
-
-
-        return view('patient.index', ['patients' => $patients, 'search' => $searchTerm]);
     }
 }
