+++ resources/views/doctor/edit.blade.php
@@ -9,7 +9,6 @@
         </div>
     </x-slot>
     <x-flash-message />
-
     <div class="py-12">
         <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
             <div class="bg-white dark:bg-gray-800 overflow-hidden shadow-sm sm:rounded-lg">
