+++ resources/views/doctor/index.blade.php
@@ -4,13 +4,11 @@
             <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">
                 Doctor Information
             </h2>
-            <form action="{{ route('doctor.search') }}" method="GET" class="flex items-center space-x-4">
+            <form method="GET" action="{{ route('doctor.index') }}" class="flex items-center space-x-4">
                 @csrf
-                <input id="search"
-                    class="w-80 border border-gray-300 rounded-md shadow-sm focus:border-blue-500 focus:ring focus:ring-blue-200 focus:ring-opacity-50"
-                    type="search" name="search" value="{{ isset($search) ? $search : '' }}" placeholder="Search Doctor"
-                    autocomplete="off" />
-                <x-input-error :messages="$errors->get('search')" class="mt-2 text-red-600" />
+                <input type="search" name="search" value="{{ request('search') }}" placeholder="Search doctors..."
+                    id="search"
+                    class="w-80 border border-gray-300 rounded-md shadow-sm focus:border-blue-500 focus:ring focus:ring-blue-200 focus:ring-opacity-50">
             </form>
         </div>
     </x-slot>
@@ -100,7 +98,7 @@
                         </tbody>
                     </table>
                     <div>
-                        {{ $doctors->links() }}
+                        {{ $doctors->appends(['search' => request('search')])->links() }}
                     </div>
                 </div>
             </div>
